+++ src/quacc/__init__.py
@@ -1,12 +1,15 @@
 """Init data for quacc"""
 
 import os
+from importlib.metadata import version
 
 from ase import Atoms
 from ase.io.jsonio import decode, encode
 
-from quacc._version import __version__
 from quacc.settings import QuaccSettings
+from quacc.util.wflows import flow, job, subflow
+
+__all__ = ["flow", "job", "subflow"]
 
 
 def atoms_as_dict(s):
@@ -25,13 +28,12 @@ def atoms_from_dict(d):
     return decode(d["atoms_json"])
 
 
+# Load the version
+__version__ = version("quacc")
+
 # Make Atoms MSONable
 Atoms.as_dict = atoms_as_dict
 Atoms.from_dict = atoms_from_dict
 
 # Load the settings
 SETTINGS = QuaccSettings()
-
-for f in {SETTINGS.SCRATCH_DIR, SETTINGS.RESULTS_DIR}:
-    if f:
-        os.makedirs(f, exist_ok=True)