+++ src/quacc/recipes/gaussian/core.py
@@ -2,20 +2,27 @@
 from __future__ import annotations
 
 import multiprocessing
+from typing import TYPE_CHECKING
 
-import covalent as ct
-from ase import Atoms
 from ase.calculators.gaussian import Gaussian
 
-from quacc.schemas.cclib import cclibSchema, summarize_run
+from quacc import job
+from quacc.schemas.atoms import fetch_atoms
+from quacc.schemas.cclib import summarize_run
+from quacc.util.atoms import get_charge, get_multiplicity
 from quacc.util.calc import run_calc
 from quacc.util.dicts import remove_dict_empties
+
+if TYPE_CHECKING:
+    from ase import Atoms
+
+    from quacc.schemas.cclib import cclibSchema
 
 LOG_FILE = f"{Gaussian().label}.log"
 GEOM_FILE = LOG_FILE
 
 
-@ct.electron
+@job
 def static_job(
     atoms: Atoms | dict,
     charge: int | None = None,
@@ -52,15 +59,8 @@ def static_job(
     RunSchema
         Dictionary of results from `quacc.schemas.cclib.summarize_run`
     """
-    atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
+    atoms = fetch_atoms(atoms)
     calc_swaps = calc_swaps or {}
-
-    charge = int(atoms.get_initial_charges().sum()) if charge is None else charge
-    multiplicity = (
-        int(1 + atoms.get_initial_magnetic_moments().sum())
-        if multiplicity is None
-        else multiplicity
-    )
 
     defaults = {
         "mem": "16GB",
@@ -68,8 +68,8 @@ def static_job(
         "nprocshared": multiprocessing.cpu_count(),
         "xc": xc,
         "basis": basis,
-        "charge": charge,
-        "mult": multiplicity,
+        "charge": get_charge(atoms) if charge is None else charge,
+        "mult": get_multiplicity(atoms) if multiplicity is None else multiplicity,
         "sp": "",
         "scf": ["maxcycle=250", "xqc"],
         "integral": "ultrafine",
@@ -90,7 +90,7 @@ def static_job(
     )
 
 
-@ct.electron
+@job
 def relax_job(
     atoms: Atoms,
     charge: int | None = None,
@@ -130,15 +130,8 @@ def relax_job(
     RunSchema
         Dictionary of results from `quacc.schemas.cclib.summarize_run`
     """
-    atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
+    atoms = fetch_atoms(atoms)
     calc_swaps = calc_swaps or {}
-
-    charge = int(atoms.get_initial_charges().sum()) if charge is None else charge
-    multiplicity = (
-        int(1 + atoms.get_initial_magnetic_moments().sum())
-        if multiplicity is None
-        else multiplicity
-    )
 
     defaults = {
         "mem": "16GB",
@@ -146,8 +139,8 @@ def relax_job(
         "nprocshared": multiprocessing.cpu_count(),
         "xc": xc,
         "basis": basis,
-        "charge": charge,
-        "mult": multiplicity,
+        "charge": get_charge(atoms) if charge is None else charge,
+        "mult": get_multiplicity(atoms) if multiplicity is None else multiplicity,
         "opt": "",
         "pop": "CM5",
         "scf": ["maxcycle=250", "xqc"],