+++ src/quacc/recipes/orca/core.py
@@ -3,21 +3,28 @@ from __future__ import annotations
 
 import multiprocessing
 import os
+from typing import TYPE_CHECKING
 
-import covalent as ct
-from ase import Atoms
 from ase.calculators.orca import ORCA, OrcaProfile
 
-from quacc import SETTINGS
-from quacc.schemas.cclib import cclibSchema, summarize_run
+from quacc import SETTINGS, job
+from quacc.schemas.atoms import fetch_atoms
+from quacc.schemas.cclib import summarize_run
+from quacc.util.atoms import get_charge, get_multiplicity
 from quacc.util.calc import run_calc
 from quacc.util.dicts import remove_dict_empties
+
+if TYPE_CHECKING:
+    from ase import Atoms
+
+    from quacc.schemas.cclib import cclibSchema
+
 
 LOG_FILE = f"{ORCA().name}.out"
 GEOM_FILE = f"{ORCA().name}.xyz"
 
 
-@ct.electron
+@job
 def static_job(
     atoms: Atoms | dict,
     charge: int | None = None,
@@ -61,7 +68,7 @@ def static_job(
     cclibSchema
         Dictionary of results from quacc.schemas.cclib.summarize_run
     """
-    atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
+    atoms = fetch_atoms(atoms)
     input_swaps = input_swaps or {}
     block_swaps = block_swaps or {}
 
@@ -86,17 +93,10 @@ def static_job(
     orcasimpleinput = " ".join(list(inputs.keys()))
     orcablocks = " ".join(list(blocks.keys()))
 
-    charge = int(atoms.get_initial_charges().sum()) if charge is None else charge
-    multiplicity = (
-        int(1 + atoms.get_initial_magnetic_moments().sum())
-        if multiplicity is None
-        else multiplicity
-    )
-
     atoms.calc = ORCA(
         profile=OrcaProfile([SETTINGS.ORCA_CMD]),
-        charge=charge,
-        mult=multiplicity,
+        charge=get_charge(atoms) if charge is None else charge,
+        mult=get_multiplicity(atoms) if multiplicity is None else multiplicity,
         orcasimpleinput=orcasimpleinput,
         orcablocks=orcablocks,
     )
@@ -109,7 +109,7 @@ def static_job(
     )
 
 
-@ct.electron
+@job
 def relax_job(
     atoms: Atoms | dict,
     charge: int | None = None,
@@ -156,7 +156,7 @@ def relax_job(
     cclibSchema
         Dictionary of results from quacc.schemas.cclib.summarize_run
     """
-    atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
+    atoms = fetch_atoms(atoms)
     input_swaps = input_swaps or {}
     block_swaps = block_swaps or {}
 
@@ -182,17 +182,10 @@ def relax_job(
     orcasimpleinput = " ".join(list(inputs.keys()))
     orcablocks = " ".join(list(blocks.keys()))
 
-    charge = int(atoms.get_initial_charges().sum()) if charge is None else charge
-    multiplicity = (
-        int(1 + atoms.get_initial_magnetic_moments().sum())
-        if multiplicity is None
-        else multiplicity
-    )
-
     atoms.calc = ORCA(
         profile=OrcaProfile([SETTINGS.ORCA_CMD]),
-        charge=charge,
-        mult=multiplicity,
+        charge=get_charge(atoms) if charge is None else charge,
+        mult=get_multiplicity(atoms) if multiplicity is None else multiplicity,
         orcasimpleinput=orcasimpleinput,
         orcablocks=orcablocks,
     )